+++ src/LASRcore/pipeline.cpp
@@ -198,6 +198,9 @@ bool Pipeline::run_loaded()
 
     if (verbose) print("Stage: %s\n", stage->get_name().c_str());
 
+    // Some stages are capable of breaking the pipeline if a conditional statement is met
+    if (stage->break_pipeline()) break;
+
     // Some stages need no input, they are connected to another stage
     // (such as pit_fill which is connected to a raster stage and does not need any point)
     success = stage->process();