+++ tests/testthat/test-pipelines.R
@@ -77,9 +77,11 @@ test_that("pipleline info works",
   expect_error(lasR:::get_pipeline_info(pipeline), "Unsupported stage: plop")
 })
 
-test_that("processor fails without reader",
+test_that("processor does not fails without reader",
 {
   f <- system.file("extdata", "bcts/", package="lasR")
 
-  expect_error(exec(hulls() + reader_las(), on = f),  "The reader must alway be the first stage of the pipeline")
+  #expect_error(exec(hulls(), on = f), "The pipeline must have a readers stage")
+  expect_error(exec(local_maximum(10) + reader_las(f), on = f),  "not preceded by a reader stage")
+  expect_error(exec(hulls() + reader_las(f), on = f),  "A 'reader' stage is missing or is at an incorrect position in the pipeline")
 })